feat: MongoDB configuration for multiple replicas

+++ infrastructure/modules/onpremise-storage/mongodb/configmap.tf
@@ -21,7 +21,9 @@ EOF
 rs.initiate({
   _id :  "rs0",
   members: [
-    { _id:  0, host:  "${local.mongodb_dns}:${local.mongodb_port}" },
+%{for i, service in kubernetes_service.mongodb~}
+    { _id:  ${i}, host:  "${service.metadata.0.name}.${service.metadata.0.namespace}:${service.spec.0.type == "NodePort" ? service.spec.0.port.0.node_port : service.spec.0.port.0.port}" },
+%{endfor~}
   ]
 })
 
@@ -31,8 +33,11 @@ EOF
 
 # TODO: put this file at another place
 CLUSTER_KEY=/data/db/cluster.key
-cp /cluster/cluster.key $CLUSTER_KEY
-chmod 400 $CLUSTER_KEY
+
+if [ ! -e $CLUSTER_KEY ] ; then
+  cp /cluster/cluster.key $CLUSTER_KEY
+  chmod 400 $CLUSTER_KEY
+fi
 
 /usr/local/bin/docker-entrypoint.sh mongod \
   --dbpath=/data/db \
@@ -46,16 +51,20 @@ chmod 400 $CLUSTER_KEY
   --noscripting \
   --replSet=rs0 &
 
-sleep 10
+sleep 15
 
-mongosh \
-  --username ${random_string.mongodb_admin_user.result} \
-  --password ${random_password.mongodb_admin_password.result} \
-  --tlsCAFile /mongodb/chain.pem \
-  --tlsAllowInvalidHostnames \
-  --tlsAllowInvalidCertificates \
-  --tls \
-  localhost:27017/admin /start/initreplica.js
+if [ "$1" == "0" ] ; then
+  mongosh \
+    --username ${random_string.mongodb_admin_user.result} \
+    --password ${random_password.mongodb_admin_password.result} \
+    --tlsCAFile /mongodb/chain.pem \
+    --tlsAllowInvalidHostnames \
+    --tlsAllowInvalidCertificates \
+    --tls \
+    localhost:27017/admin /start/initreplica.js
+fi
+
+
 
 wait
 
@@ -80,7 +89,7 @@ resource "kubernetes_config_map" "mongo_start_sh" {
     namespace = var.namespace
   }
   data = {
-    "mongostart.sh" = local.mongo_start_sh
+    "mongostart.sh"  = local.mongo_start_sh
     "initreplica.js" = local.init_replica_js
   }
 }
@@ -88,4 +97,9 @@ resource "kubernetes_config_map" "mongo_start_sh" {
 resource "local_file" "mongodb_js_file" {
   content  = local.mongodb_js
   filename = "${path.root}/generated/configmaps/mongodb/mongodb.js"
+}
+
+resource "local_file" "init_replica_js" {
+  content  = local.init_replica_js
+  filename = "${path.root}/generated/configmaps/mongodb/initreplica.js"
 }
